refactor(glyph): split the raw-mask interpolation lexer into its own module
`raw_mask.rs` grew past the repo's 350-line source limit, so the
interpolation expression lexer and the mask tests move into sibling
`raw_mask/interpolation.rs` and `raw_mask/tests.rs` modules. No behaviour
change.
