+++ crates/vize_glyph/src/formatter/raw_mask.rs
@@ -1,3 +1,9 @@
+mod interpolation;
+#[cfg(test)]
+mod tests;
+
+use interpolation::InterpolationScan;
+
 /// Per-line "this line is inside a whitespace-significant block" mask.
 ///
 /// Lines inside `<pre>`, `<textarea>`, `v-pre`, multi-line comments,
@@ -150,120 +156,6 @@ pub(super) fn compute_raw_line_mask(lines: &[&[u8]]) -> Vec<bool> {
     mask
 }
 
-/// Lexer state for the JS expression inside a `{{ … }}` interpolation.
-///
-/// `}}` and backticks are only structural in code position: inside a string,
-/// a comment, or a nested template literal they are ordinary characters. A
-/// scan that ignored that let `{{ "}}" + ` … ` }}` close the interpolation at
-/// the string's `}}`, so the literal's lines lost their raw marking and their
-/// runtime value was reformatted, and a single boolean could not represent a
-/// template literal nested inside a `${ … }` substitution.
-///
-/// This mirrors `template_literal_quasi_line_starts` (#3247), which decides
-/// the same question for the formatted expression, so both layers agree on
-/// which lines are quasi text. Regex literals are not modelled; a backtick
-/// inside one is rare and would at worst leave a line indented.
-#[derive(Default)]
-struct InterpolationScan {
-    /// Set by `{{`, cleared by the matching `}}`.
-    active: bool,
-    /// Nested template-literal / `${ … }` frames within the expression.
-    frames: Vec<ExprFrame>,
-    /// Open `'…'` / `"…"` string, if any. Reset at every line boundary,
-    /// since such strings cannot span a newline.
-    string: Option<u8>,
-    /// Inside a `/* … */` comment.
-    in_block_comment: bool,
-}
-
-enum ExprFrame {
-    /// Inside backticks, in quasi text (part of the string's runtime value).
-    TemplateLiteral,
-    /// Inside `${ … }`; tracks `{`/`}` nesting within the substitution.
-    Substitution(u32),
-}
-
-impl InterpolationScan {
-    /// Whether the current line's first byte lies in template-literal quasi
-    /// text, which the SFC layer must leave unindented.
-    fn line_starts_in_quasi(&self) -> bool {
-        self.string.is_none()
-            && !self.in_block_comment
-            && matches!(self.frames.last(), Some(ExprFrame::TemplateLiteral))
-    }
-
-    /// Consume one token at `cursor`, returning the next cursor position.
-    fn step(&mut self, bytes: &[u8], cursor: usize) -> usize {
-        let byte = bytes[cursor];
-        if self.in_block_comment {
-            if bytes[cursor..].starts_with(b"*/") {
-                self.in_block_comment = false;
-                return cursor + 2;
-            }
-            return cursor + 1;
-        }
-        if let Some(quote) = self.string {
-            if byte == b'\\' {
-                return cursor + 2;
-            }
-            if byte == quote {
-                self.string = None;
-            }
-            return cursor + 1;
-        }
-        if matches!(self.frames.last(), Some(ExprFrame::TemplateLiteral)) {
-            return match byte {
-                b'\\' => cursor + 2,
-                b'`' => {
-                    self.frames.pop();
-                    cursor + 1
-                }
-                b'$' if bytes.get(cursor + 1) == Some(&b'{') => {
-                    self.frames.push(ExprFrame::Substitution(0));
-                    cursor + 2
-                }
-                _ => cursor + 1,
-            };
-        }
-        match byte {
-            b'/' if bytes.get(cursor + 1) == Some(&b'/') => bytes.len(),
-            b'/' if bytes.get(cursor + 1) == Some(&b'*') => {
-                self.in_block_comment = true;
-                cursor + 2
-            }
-            b'\'' | b'"' => {
-                self.string = Some(byte);
-                cursor + 1
-            }
-            b'`' => {
-                self.frames.push(ExprFrame::TemplateLiteral);
-                cursor + 1
-            }
-            b'{' => {
-                if let Some(ExprFrame::Substitution(depth)) = self.frames.last_mut() {
-                    *depth += 1;
-                }
-                cursor + 1
-            }
-            b'}' if self.frames.is_empty() && bytes[cursor..].starts_with(b"}}") => {
-                self.active = false;
-                cursor + 2
-            }
-            b'}' => {
-                match self.frames.last_mut() {
-                    Some(ExprFrame::Substitution(0)) => {
-                        self.frames.pop();
-                    }
-                    Some(ExprFrame::Substitution(depth)) => *depth -= 1,
-                    _ => {}
-                }
-                cursor + 1
-            }
-            _ => cursor + 1,
-        }
-    }
-}
-
 fn literal_attr_quote(line: &[u8], quote_pos: usize) -> bool {
     attr_name_before_quote(line, quote_pos).is_none_or(|name| {
         !directive_expr_attr(name) || verbatim_multiline_directive_attr(name, line, quote_pos)
@@ -361,78 +253,4 @@ fn starts_with_ascii_ci(haystack: &[u8], needle: &[u8]) -> bool {
             .iter()
             .zip(needle.iter())
             .all(|(a, b)| a.eq_ignore_ascii_case(b))
-}
-
-#[cfg(test)]
-mod interpolation_literal_tests {
-    use super::compute_raw_line_mask;
-
-    fn mask(source: &str) -> Vec<bool> {
-        let lines: Vec<&[u8]> = source.lines().map(|l| l.as_bytes()).collect();
-        compute_raw_line_mask(&lines)
-    }
-
-    #[test]
-    fn lines_inside_an_interpolation_template_literal_are_raw() {
-        // Every byte between the backticks is part of the string's runtime
-        // value, so the SFC layer must not indent these lines (#3334).
-        let source =
-            "<div>\n  {{ items.map((p) => `\n${p} {\n  --a: b;\n}\n`).join(\"\") }}\n</div>";
-        // Lines 2..=5 all *begin* inside the literal — including the line
-        // that closes it, whose leading bytes would otherwise be indented
-        // into the string's value.
-        assert_eq!(mask(source), [false, false, true, true, true, true, false]);
-    }
-
-    #[test]
-    fn ordinary_interpolation_lines_stay_indentable() {
-        let source = "<div>\n  {{ a\n    + b }}\n</div>";
-        assert_eq!(mask(source), [false, false, false, false]);
-    }
-
-    #[test]
-    fn a_closing_brace_pair_inside_a_string_does_not_end_the_interpolation() {
-        // The `}}` lives in a JS string, so the literal that follows it is
-        // still inside the interpolation and its lines stay raw.
-        let source = "<div>\n  {{ \"}}\" + `\nfoo\n` }}\n</div>";
-        assert_eq!(mask(source), [false, false, true, true, false]);
-    }
-
-    #[test]
-    fn a_template_literal_nested_in_a_substitution_is_tracked() {
-        let source = "<div>\n  {{ `${ xs.map((x) => `\na\n`) }\nb\n` }}\n</div>";
-        assert_eq!(mask(source), [false, false, true, true, true, true, false]);
-    }
-
-    #[test]
-    fn a_backtick_inside_a_comment_is_not_structural() {
-        let source = "<div>\n  {{ /* ` */ a }}\n  <span>y</span>\n</div>";
-        assert_eq!(mask(source), [false, false, false, false]);
-    }
-
-    #[test]
-    fn an_unterminated_marker_stays_text_and_keeps_markup_tracking() {
-        // Vue treats a `{{` with no `}}` as text, so the scanner must not
-        // enter expression mode and lose the `<pre>` region behind it.
-        let source = "<div>\n  {{ mustache syntax, unclosed\n  <pre>\nx\n  </pre>\n</div>";
-        assert_eq!(
-            mask(source),
-            [false, false, false, true, true, false],
-            "an unmatched `{{{{` must not disable markup scanning"
-        );
-    }
-
-    #[test]
-    fn an_unbalanced_quote_does_not_swallow_later_lines() {
-        // `'a) }}` leaves a quote open at the line end; strings cannot span a
-        // newline, so the literal on the following lines is still tracked.
-        let source = "<div>\n  {{ t('a) }}\n  {{ `\nx\n` }}\n</div>";
-        assert_eq!(mask(source), [false, false, false, true, true, false]);
-    }
-
-    #[test]
-    fn a_closed_literal_does_not_leak_into_later_lines() {
-        let source = "<div>\n  {{ `x` }}\n  <span>y</span>\n</div>";
-        assert_eq!(mask(source), [false, false, false, false]);
-    }
 }
